// scripts/CMakePresetGenerator/src/schema.d.ts
export type Schema =
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 1;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV1;
    }
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 2;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV1;
      buildPresets?: BuildPresetsV2;
      testPresets?: TestPresetsV2;
    }
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 3;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV3;
      buildPresets?: BuildPresetsV3;
      testPresets?: TestPresetsV3;
    }
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 4;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV3;
      buildPresets?: BuildPresetsV4;
      testPresets?: TestPresetsV3;
      include?: Include;
    }
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 5;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV3;
      buildPresets?: BuildPresetsV4;
      testPresets?: TestPresetsV5;
      include?: Include;
    }
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 6;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV3;
      buildPresets?: BuildPresetsV4;
      testPresets?: TestPresetsV6;
      packagePresets?: PackagePresetsV6;
      workflowPresets?: WorkflowPresetsV6;
      include?: Include;
    }
  | {
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 7;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV7;
      buildPresets?: BuildPresetsV4;
      testPresets?: TestPresetsV6;
      packagePresets?: PackagePresetsV6;
      workflowPresets?: WorkflowPresetsV6;
      include?: Include;
    }
  | {
      $schema?: $Schema;
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 8;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV7;
      buildPresets?: BuildPresetsV4;
      testPresets?: TestPresetsV6;
      packagePresets?: PackagePresetsV6;
      workflowPresets?: WorkflowPresetsV6;
      include?: Include;
    }
  | {
      $schema?: $Schema;
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 9;
      cmakeMinimumRequired?: CmakeMinimumRequiredV1;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV7;
      buildPresets?: BuildPresetsV4;
      testPresets?: TestPresetsV6;
      packagePresets?: PackagePresetsV6;
      workflowPresets?: WorkflowPresetsV6;
      include?: Include;
    }
  | {
      $schema?: $Schema;
      $comment?: $Comment;
      /**
       * A required integer representing the version of the JSON schema.
       */
      version?: 10;
      cmakeMinimumRequired?: CmakeMinimumRequiredV10;
      vendor?: Vendor;
      configurePresets?: ConfigurePresetsV10;
      buildPresets?: BuildPresetsV10;
      testPresets?: TestPresetsV10;
      packagePresets?: PackagePresetsV10;
      workflowPresets?: WorkflowPresetsV10;
      include?: Include;
    };
/**
 * An optional object representing the minimum version of CMake needed to build this project. Available in version 1 and higher.
 */
export type CmakeMinimumRequiredV1 = CmakeMinimumRequiredPropertiesV1 & {
  major?: unknown;
  minor?: unknown;
  patch?: unknown;
};
/**
 * An optional array of configure preset objects. Available in version 1 and higher.
 */
export type ConfigurePresetsV1 = ConfigurePresetsItemsV1 & ConfigurePresetsV11;
// ...
export type ConfigurePresetsV11 = {
  name: unknown;
  hidden?: unknown;
  inherits?: unknown;
  vendor?: unknown;
  displayName?: unknown;
  description?: unknown;
  generator?: unknown;
  architecture?: ConfigurePresetsArchitectureV1;
  toolset?: ConfigurePresetsToolsetV1;
  binaryDir?: unknown;
  cmakeExecutable?: unknown;
  cacheVariables?: {
    [k: string]: ConfigurePresetsCacheVariablesAdditionalPropertiesV1;
  };
  environment?: unknown;
  warnings?: {
    dev?: unknown;
    deprecated?: unknown;
    uninitialized?: unknown;
    unusedCli?: unknown;
    systemVars?: unknown;
  };
  errors?: {
    dev?: unknown;
    deprecated?: unknown;
  };
  debug?: {
    output?: unknown;
    tryCompile?: unknown;
    find?: unknown;
  };
}[];
/**
 * An optional array of build preset objects. Used to specify arguments to cmake --build. Available in version 2 and higher.
 */
export type BuildPresetsV2 = BuildPresetsItemsV2 & BuildPresetsV21;
/**
 * An optional array of build preset objects. Used to specify arguments to cmake --build. Available in version 2 and higher.
 */
export type BuildPresetsItemsV2 = {
  /**
   * A required string representing the machine-friendly name of the preset. This identifier is used in the --preset argument. There must not be two presets (configure, build, test, package, or workflow) in the union of CMakePresets.json and CMakeUserPresets.json in the same directory with the same name.
   */
  name?: string;
  /**
   * An optional boolean specifying whether or not a preset should be hidden. If a preset is hidden, it cannot be used in the --preset argument and does not have to have a valid configurePreset, even from inheritance. Hidden presets are intended to be used as a base for other presets to inherit via the inherits field.
   */
  hidden?: boolean;
  inherits?: string | string[];
  /**
   * An optional string specifying the name of a configure preset to associate with this build preset. If configurePreset is not specified, it must be inherited from the inherits preset (unless this preset is hidden). The build tree directory is inferred from the configure preset.
   */
  configurePreset?: string;
  /**
   * An optional map containing vendor-specific information. CMake does not interpret the contents of this field except to verify that it is a map if it does exist. However, it should follow the same conventions as the root-level vendor field. If vendors use their own per-preset vendor field, they should implement inheritance in a sensible manner when appropriate.
   */
  vendor?: {
    [k: string]: unknown;
  };
  /**
   * An optional string with a human-friendly name of the preset.
   */
  displayName?: string;
  /**
   * An optional string with a human-friendly description of the preset.
   */
  description?: string;
  /**
   * An optional boolean that defaults to true. If true, the environment variables from the associated configure preset are inherited after all inherited build preset environments, but before environment variables explicitly specified in this build preset.
   */
  inheritConfigureEnvironment?: boolean;
  /**
   * An optional map of environment variables. The key is the variable name (which must not be an empty string). Each variable is set regardless of whether or not a value was given to it by the process's environment. This field supports macro expansion, and environment variables in this map may reference each other, and may be listed in any order, as long as such references do not cause a cycle (for example,if ENV_1 is $env{ENV_2}, ENV_2 may not be $env{ENV_1}.) Environment variables are inherited through the inherits field, and the preset's environment will be the union of its own environment and the environment from all its parents. If multiple presets in this union define the same variable, the standard rules of inherits are applied. Setting a variable to null causes it to not be set, even if a value was inherited from another preset.
   */
  environment?: {
    [k: string]: null | string;
  };
  /**
   * An optional integer. Equivalent to passing --parallel or -j on the command line.
   */
  jobs?: number;
  targets?: string | string[];
  /**
   * An optional string. Equivalent to passing --config on the command line.
   */
  configuration?: string;
  /**
   * An optional boolean. If true, equivalent to passing --clean-first on the command line.
   */
  cleanFirst?: boolean;
  /**
   * An optional boolean. If true, equivalent to passing --verbose on the command line.
   */
  verbose?: boolean;
  /**
   * An optional array of strings. Equivalent to passing options after -- on the command line.
   */
  nativeToolOptions?: string[];
  [k: string]: unknown;
}[];
export type BuildPresetsV21 = {
  name: unknown;
  hidden?: unknown;
  inherits?: unknown;
  configurePreset?: unknown;
  vendor?: unknown;
  displayName?: unknown;
  description?: unknown;
  inheritConfigureEnvironment?: unknown;
  environment?: unknown;
  jobs?: unknown;
  targets?: unknown;
  configuration?: unknown;
  cleanFirst?: unknown;
  verbose?: unknown;
  nativeToolOptions?: unknown;
}[];
/**
 * An optional array of test preset objects. Used to specify arguments to ctest. Available in version 2 and higher.
 */
export type TestPresetsV2 = TestPresetsItemsV2 & TestPresetsV21;
// ...
export type ConfigurePresetsV3 = ConfigurePresetsItemsV3 & ConfigurePresetsItemsV1 & ConfigurePresetsV31;
// ...
export type BuildPresetsV3 = BuildPresetsItemsV3 & BuildPresetsItemsV2 & BuildPresetsV31;
/**
 * An optional array of build preset objects. Used to specify arguments to cmake --build. Available in version 3 and higher.
 */
export type BuildPresetsItemsV3 = unknown[];
export type BuildPresetsV31 = {
  name: unknown;
  hidden?: unknown;
  inherits?: unknown;
  configurePreset?: unknown;
  vendor?: unknown;
  displayName?: unknown;
  description?: unknown;
  inheritConfigureEnvironment?: unknown;
  environment?: unknown;
  jobs?: unknown;
  targets?: unknown;
  configuration?: unknown;
  cleanFirst?: unknown;
  verbose?: unknown;
  nativeToolOptions?: unknown;
  condition?: TopConditionV3;
}[];
/**
 * An optional array of test preset objects. Used to specify arguments to ctest. Available in version 3 and higher.
 */
export type TestPresetsV3 = TestPresetsItemsV3 & TestPresetsItemsV2 & TestPresetsV31;
/**
 * An optional array of test preset objects. Used to specify arguments to ctest. Available in version 3 and higher.
 */
export type TestPresetsItemsV3 = unknown[];
// ...
export type BuildPresetsV4 = BuildPresetsItemsV4 & BuildPresetsItemsV3 & BuildPresetsItemsV2 & BuildPresetsV41;
// ...
export type Include = string[];
/**
 * An optional array of test preset objects. Used to specify arguments to ctest. Available in version 5 and higher.
 */
export type TestPresetsV5 = TestPresetsItemsV5 & TestPresetsItemsV3 & TestPresetsItemsV2 & TestPresetsV51;
// ...
export type TestPresetsV6 = TestPresetsItemsV6 &
  TestPresetsItemsV5 &
  TestPresetsItemsV3 &
  TestPresetsItemsV2 &
  TestPresetsV61;
// ...
export type PackagePresetsV6 = PackagePresetsItemsV6 & PackagePresetsV61;
// ...
export type WorkflowPresetsV6 = WorkflowPresetsItemsV6 & WorkflowPresetsV61;
// ...
export type ConfigurePresetsV7 = ConfigurePresetsItemsV7 &
  ConfigurePresetsItemsV3 &
  ConfigurePresetsItemsV1 &
  ConfigurePresetsV71;
// ...
export type $Schema = string;
export type $Comment = string | [string, ...string[]];
/**
 * An optional object representing the minimum version of CMake needed to build this project. Available in version 10 and higher.
 */
export type CmakeMinimumRequiredV10 = CmakeMinimumRequiredPropertiesV10 &
  CmakeMinimumRequiredPropertiesV1 & {
    $comment?: unknown;
    major?: unknown;
    minor?: unknown;
    patch?: unknown;
  };
/**
 * An optional array of configure preset objects. Available in version 10 and higher.
 */
export type ConfigurePresetsV10 = ConfigurePresetsItemsV10 &
  ConfigurePresetsItemsV7 &
  ConfigurePresetsItemsV3 &
  ConfigurePresetsItemsV1 &
  ConfigurePresetsV101;
// ...
export type BuildPresetsV10 = BuildPresetsItemsV10 &
  BuildPresetsItemsV4 &
  BuildPresetsItemsV3 &
  BuildPresetsItemsV2 &
  BuildPresetsV101;
/**
 * An optional array of build preset objects. Used to specify arguments to cmake --build. Available in version 10 and higher.
 */
export type BuildPresetsItemsV10 = {
  $comment?: $Comment;
  [k: string]: unknown;
}[];
// ...
export type TestPresetsV10 = TestPresetsItemsV10 &
  TestPresetsItemsV6 &
  TestPresetsItemsV5 &
  TestPresetsItemsV3 &
  TestPresetsItemsV2 &
  TestPresetsV101;
// ...
export type PackagePresetsV10 = PackagePresetsItemsV10 & PackagePresetsItemsV6 & PackagePresetsV101;
// ...
export type WorkflowPresetsV10 = WorkflowPresetsItemsV10 & WorkflowPresetsItemsV6 & WorkflowPresetsV101;
// ...
export interface CmakeMinimumRequiredPropertiesV1 {
  /**
   * An optional integer representing the major version.
   */
  major?: number;
  /**
   * An optional integer representing the minor version.
   */
  minor?: number;
  /**
   * An optional integer representing the patch version.
   */
  patch?: number;
  [k: string]: unknown;
}
// ...
export interface Vendor {
  [k: string]: unknown;
}
// ...
export interface CmakeMinimumRequiredPropertiesV10 {
  $comment?: $Comment;
  [k: string]: unknown;
}